pridane funkcionality do adminskej casti - moznost editacie filmu a vyskocenie okna editacia s doterajsimi udajmi - podobne pri pridavani noveho filmu bez prednastavenych hodnot

+++ App/Controllers/AdminController.php
@@ -13,27 +13,81 @@ class AdminController extends AControllerBase
 
     public function index()
     {
-        return $this->html(Film::getAll());
+        $user = $this->app->getAuthController()->getUser();
+        if ($user != null) {
+            if ($user->getUserType() == 2) {
+                return $this->html(Film::getAll());
+            } else {
+                $notlogged = array('Not authorized');
+                return $this->json($notlogged);
+            }
+        } else {
+            $notlogged = array('Not logged');
+            return $this->json($notlogged);
+        }
     }
 
     public function edit()
     {
-        // TODO: Implement index() method.
+        $user = $this->app->getAuthController()->getUser();
+        if ($user != null) {
+            if ($user->getUserType() == 2) {
+                $form_data = $this->app->getRequest()->getPost();
+                $film = Film::getOne($form_data['id_film']);
+                $film->setImg($form_data['img']);
+                $film->setTitle($form_data['title']);
+                $film->setAboutFilm($form_data['about_film']);
+                $film->setFilmType($form_data['film_type']);
+                $film->save();
+            } else {
+                $notlogged = array('Not authorized');
+                return $this->json($notlogged);
+            }
+
+        } else {
+            $notlogged = array('Not logged');
+            return $this->json($notlogged);
+        }
     }
 
     public function add()
     {
+        $user = $this->app->getAuthController()->getUser();
+        if ($user != null) {
+            if ($user->getUserType() == 2) {
+                $form_data = $this->app->getRequest()->getPost();
+                $film = new Film($form_data['img'], $form_data['title'], $form_data['about_film'], $form_data['film_type']);
+                $film->save();
+            } else {
+                $notlogged = array('Not authorized');
+                return $this->json($notlogged);
+            }
+        } else {
+            $notlogged = array('Not logged');
+            return $this->json($notlogged);
+        }
     }
 
     public function delete()
     {
-        $form_data = $this->app->getRequest()->getPost();
-        $favfilms = FavFilm::getAll("id_film =" . $form_data['id_film']);
-        foreach ($favfilms as $fav) {
-            $fav->delete("id_film=" . $form_data['id_film']);
-            break;
+        $user = $this->app->getAuthController()->getUser();
+        if ($user != null) {
+            if ($user->getUserType() == 2) {
+                $form_data = $this->app->getRequest()->getPost();
+                $favfilms = FavFilm::getAll("id_film =" . $form_data['id_film']);
+                foreach ($favfilms as $fav) {
+                    $fav->delete("id_film=" . $form_data['id_film']);
+                    break;
+                }
+                $film = Film::getOne($form_data['id_film']);
+                $film->delete();
+            } else {
+                $notlogged = array('Not authorized');
+                return $this->json($notlogged);
+            }
+        } else {
+            $notlogged = array('Not logged');
+            return $this->json($notlogged);
         }
-        $film = Film::getOne($form_data['id_film']);
-        $film->delete();
     }
 }
